% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codon_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Codon table - DNA to amino acid}
\usage{
dna_codon_to_aa(dna = aa)
}
\description{
dna triplet codons and the amino acid it codes for
}
\details{
R/dna_codons_to_aa.R
}
